requirements.txt
aiofiles==24.1.0
aiogram==2.23.1
aiohappyeyeballs==2.4.2
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
Babel==2.9.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.8.30
charset-normalizer==3.3.2
convertdate==2.4.0
Django==5.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
exceptiongroup==1.2.2
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.10
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
Khayyam==3.0.17
magic-filter==1.0.12
multidict==6.1.0
packaging==24.1
persiantools==4.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
PyMeeus==0.5.12
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
rpds-py==0.20.0
shortuuid==1.0.13
sniffio==1.3.1
soupsieve==2.5
sqlparse==0.5.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
yarl==1.13.1
